% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Downscale BoM weather and forecasts to hourly data for spatial interpolation}
\usage{
downscale()
}
\description{
Downscale BoM weather and forecasts to hourly data for spatial interpolation
}
\details{
An R function which formats Australian Bureau of Meteorology (BoM) hourly
weather data and Queensland's Scientific Information for Land Owners (SILO)
daily weather data for validation of temporal downscaling of SILO temperature
data to hourly values.
}
\examples{
\dontrun{downscale()}

}
