% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Interpolate}
\usage{
interpolate(downscaled_temp)
}
\arguments{
\item{downscaled_temp}{Hourly temperature data from
\code{\link{make_hourly_forecast}}}
}
\description{
Interpolate
}
\details{
Interpolates weather data
}
\examples{
\dontrun{Qld_weather <- interpolate()}

}
