% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_hourly.R
\name{interpolate_hourly}
\alias{interpolate_hourly}
\title{Interpolate BoM Précis Daily Forecast Data That Have Been Downscaled to Hourly}
\usage{
interpolate_hourly(downscaled_temp)
}
\arguments{
\item{downscaled_temp}{Hourly temperature data from
\code{\link{make_hourly_forecast}}}
}
\description{
Interpolate BoM Précis Daily Forecast Data That Have Been Downscaled to Hourly
}
\details{
Interpolates forecast data that have been downscaled using the
\code{\link{make_hourly_forecast}} function and returns a list of
hourly #' surfaces for the next six-days forecast at 250x250m resolution.
List names are the Julian day followed by the hour, e.g., "316.1" being the
316th day of the year and the first hour of the 24 hours.
}
\examples{

\dontrun{Qld_weather <- interpolate_hourly(x)}

}
