% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stress_hours.R
\name{mungbean_stress}
\alias{mungbean_stress}
\title{Calculate Hours of Mungbean Heat Stress}
\usage{
mungbean_stress(interpolated_surfaces)
}
\arguments{
\item{forecast_surface}{A list of interpolated hourly temperature surfaces
from \code{\link{interpolate_hourly}}}
}
\description{
Calculate Hours of Mungbean Heat Stress
}
\details{
Calculates hours of heat stress a mungbean crop could experience given a
list of interpolated surfaces of a BoM six-day forecast from
\link{interpolate_hourly}.
}
\examples{

\dontrun{mungbean_stress <- mungbean_stress(x)}

}
