% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_test_forecast.R
\name{unit_test_forecast}
\alias{unit_test_forecast}
\title{Test BoM Précis Forecast for Changes and Send an Alert on Changes or Failure}
\usage{
unit_test_forecast(AUS_forecast)
}
\value{
E-mail if issues are encountered.
}
\description{
Test BoM Précis Forecast for Changes and Send an Alert on Changes or Failure
}
\details{
Tests the current BoM forecast against the prior day's forecast for changes
and sends an e-mail alert on change or failure.
}
\examples{
\dontrun{unit_test(AUS_forecast)}

}
\author{
Keith Pembleton
}
