\name{win.stat.Luo}
\alias{win.stat.Luo}
\title{
Win Stat Function Provided in the Supplementary Material of Luo et al. (2015).
}
\description{
A function to calculate the win ratio and the win difference provided in the supplementary material of Luo et al (2015).
}
\value{
  \item{wratio}{The win ratio.}
  \item{vwratio}{Variance for the win ratio.}
  \item{wdiff}{The win difference.}
  \item{vwdiff}{Variance for the win difference.}
}
\references{
\itemize{
    \item{Luo, X., Tian, H., Mohanty, S. and Tsai, W.Y., 2015. An alternative approach to confidence interval estimation for the win ratio statistic. Biometrics, 71(1), pp.139-145.}
  }
}
