% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WINS-package.R
\docType{package}
\name{WINS-package}
\alias{WINS}
\alias{WINS-package}
\title{WINS: Crop Stress Weather Information Delivery System}
\description{
Identifies likely heat or cold stress events in crops and send an alert via e-mail to subscribers. The chillR package is used to generate hourly weather data from Australia Bureau of Meteorology (BOM).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ToowoombaTrio/WINS}
  \item \url{https://toowoombatrio.github.io/WINS/}
  \item Report bugs at \url{https://github.com/ToowoombaTrio/WINS/issues}
}

}
\author{
\strong{Maintainer}: Adam Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID})

Authors:
\itemize{
  \item Keith Pembleton \email{keith.pembleton@usq.edu.au} (\href{https://orcid.org/0000-0002-1896-4516}{ORCID})
}

Other contributors:
\itemize{
  \item Gordon Grundy \email{gordongrundy@gmail.com} [contributor]
}

}
\keyword{internal}
