% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale.R
\name{downscale}
\alias{downscale}
\title{Downscale daily weather data to hourly}
\usage{
downscale()
}
\description{
Downscale daily weather data to hourly
}
\details{
Downscales Australian Bureau of Meteorology (BOM) daily weather forecasts and
Queensland's Scientific Information for Land Owners (SILO) daily weather data
to hourly values.
}
\examples{
\dontrun{Qld_hourly <- downscale()}

}
