% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hourly_forecast.R
\name{make_hourly_forecast}
\alias{make_hourly_forecast}
\title{Downscale BoM Précis Daily Forecast Data to Hourly Values}
\usage{
make_hourly_forecast(forecast)
}
\arguments{
\item{forecast}{A BoM generated forecast as returned by
\code{\link{get_precis_forecast}}.}
}
\value{
A list of \code{data.frame} objects sorted by Julian day for spatial
interpolation.
}
\description{
Downscale BoM Précis Daily Forecast Data to Hourly Values
}
\details{
Creates a dataframe of hourly temperature values based on the Australian
Bureau of Meteorology (BOM) daily précis weather forecasts by using
\code{\link{make_hourly_temps}} from \code{\link{chillR}}.
}
\examples{

\dontrun{Qld_hourly <- make_hourly_forecast(x)}

}
\author{
Adam H Sparks
}
