% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_test_forecast.R
\name{unit_test_forecast}
\alias{unit_test_forecast}
\title{Test BoM Precis Forecast for Changes and Send an Alert on Changes or Failure}
\usage{
unit_test_forecast(AUS_forecast, project_email_address, project_email_password,
  report_email_addresses, yesterdays_forecast)
}
\arguments{
\item{AUS_forecast}{A forecast object generated by
\code{\link{get_precis_forecast}} to check for changes.}

\item{project_email_address}{Project e-mail account used to send alerts.}

\item{project_email_password}{Password for e-mail account used to send
alerts.}

\item{report_email_addresses}{An atomic vector of e-mail addresses to which
error reports should be sent.}

\item{yesterdays_forecast}{A csv file of the prior day's forecast used as
basis for checking current day's forecast for changes}
}
\value{
E-mail if issues are encountered.
}
\description{
Test BoM Precis Forecast for Changes and Send an Alert on Changes or Failure
}
\details{
Tests the current BoM forecast against the prior day's forecast for changes
and sends an e-mail alert on change or failure.
}
\examples{
\dontrun{unit_test(forecast = AUS_forecast)}

}
\author{
Keith Pembleton
}
